Complete activity data collection
Collect stream data for latitude/longitude and HR where available. Link  the BBC Weather scraper project to be able to access weather data collected by a separate program. Also, write a function to determine if an activity should be processed by the program (skipping if false). The criteria are: activity is a run, title is default, description is empty. Outstanding progress.