Implement + validate route templates.
Route templates extend standard templates by allowing input of list of lat/long points that need to be reached, with a set margin of error to accont for continuous nature of lat/long. Similarly, an additional restriction can be set - a blacklist of points where the route is mismatched if any of the points are reached with a set margin of error. Just need to serialise templates into data classes before the Config data implementation is done! Then can implement actual naming logic using strictly validated data.